eliminate size_t vs. int warning for Win x64

+++ lib/device.c
@@ -118,7 +118,7 @@ static SOCKET server_connect(const char *host, unsigned short nport)
 		SOCKET sock;
 		int family = addr->ai_family;
 		struct sockaddr *sa = addr->ai_addr;
-		int sa_len = addr->ai_addrlen;
+		int sa_len = (int) addr->ai_addrlen; /* cast eliminates warning for (Windows) x64: size_t vs. int */
 
 #else
